requirements.txt
# SR Trading Bot — Dependencies
# pip install -r requirements.txt

# ============================================================
# MetaTrader 5 — Instalar UNA de las dos opciones:
# ============================================================

# OPCIÓN A: Windows (librería oficial de MetaQuotes)
# pip install MetaTrader5>=5.0.45

# OPCIÓN B: macOS con Docker (compatible Apple Silicon M1/M2/M3/M4)
# pip install siliconmetatrader5

# ============================================================
# Dependencias comunes (instalar siempre)
# ============================================================

# Data & Analysis
pandas>=2.0
numpy>=1.24

# Configuration
PyYAML>=6.0

# Notifications
requests>=2.31

# Backtesting & Visualization
matplotlib>=3.7
